Blue fixes
