+++ app/src/main/java/com/example/vinilosmobileapp/ui/artist/ArtistFragment.kt
@@ -1,0 +1,27 @@
+package com.example.vinilosmobileapp.ui.artist
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.vinilosmobileapp.databinding.FragmentArtistBinding
+
+class ArtistFragment : Fragment() {
+
+    private var _binding: FragmentArtistBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentArtistBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
