--- app/src/main/java/com/example/vinilosmobileapp/ui/createalbum/CreateAlbumViewModel.kt
@@ -1,17 +1,0 @@
-package com.example.vinilosmobileapp.ui.createalbum
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class CreateAlbumViewModel : ViewModel() {
-
-    private val _albumCreated = MutableLiveData<Boolean>()
-    val albumCreated: LiveData<Boolean> = _albumCreated
-
-    fun submitAlbum(name: String, year: Int, artist: String, genre: String) {
-        // Here you'd call your use case or repository
-        // Fake success callback
-        _albumCreated.value = true
-    }
-}
