+++ app/src/main/res/color/bottom_nav_color_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/purple_700" android:state_checked="true" />
+    <item android:color="#9E9E9E" />
+</selector>
