--- create_album_ui_patch.txt
@@ -1,105 +1,0 @@
-diff --git a/app/src/main/java/com/example/vinilosmobileapp/ui/album/CreateAlbumFragment.kt b/app/src/main/java/com/example/vinilosmobileapp/ui/album/CreateAlbumFragment.kt
-new file mode 100644
-index 0000000..05f16c2
---- /dev/null
-+++ b/app/src/main/java/com/example/vinilosmobileapp/ui/album/CreateAlbumFragment.kt
-@@ -0,0 +1,38 @@
-+package com.example.vinilosmobileapp.ui.album
-+
-+import android.os.Bundle
-+import android.view.LayoutInflater
-+import android.view.View
-+import android.view.ViewGroup
-+import android.widget.Toast
-+import androidx.appcompat.app.AppCompatActivity
-+import androidx.fragment.app.Fragment
-+import androidx.fragment.app.viewModels
-+import com.example.vinilosmobileapp.R
-+import com.example.vinilosmobileapp.databinding.FragmentCreateAlbumBinding
-+
-+class CreateAlbumFragment : Fragment() {
-+
-+    private var _binding: FragmentCreateAlbumBinding? = null
-+    private val binding get() = _binding!!
-+
-+    override fun onCreateView(
-+        inflater: LayoutInflater, container: ViewGroup?,
-+        savedInstanceState: Bundle?
-+    ): View {
-+        _binding = FragmentCreateAlbumBinding.inflate(inflater, container, false)
-+        return binding.root
-+    }
-+
-+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-+        super.onViewCreated(view, savedInstanceState)
-+
-+        (requireActivity() as AppCompatActivity).supportActionBar?.title =
-+            getString(R.string.create_new_album)
-+    }
-+
-+    override fun onDestroyView() {
-+        super.onDestroyView()
-+        _binding = null
-+    }
-+}
-diff --git a/app/src/main/java/com/example/vinilosmobileapp/ui/album/CreateAlbumViewModel.kt b/app/src/main/java/com/example/vinilosmobileapp/ui/album/CreateAlbumViewModel.kt
-new file mode 100644
-index 0000000..0f9d13e
---- /dev/null
-+++ b/app/src/main/java/com/example/vinilosmobileapp/ui/album/CreateAlbumViewModel.kt
-@@ -0,0 +1,17 @@
-+package com.example.vinilosmobileapp.ui.album
-+
-+import androidx.lifecycle.LiveData
-+import androidx.lifecycle.MutableLiveData
-+import androidx.lifecycle.ViewModel
-+
-+class CreateAlbumViewModel : ViewModel() {
-+
-+    private val _albumCreated = MutableLiveData<Boolean>()
-+    val albumCreated: LiveData<Boolean> = _albumCreated
-+
-+    fun submitAlbum(name: String, year: Int, artist: String, genre: String) {
-+        // Here you'd call your use case or repository
-+        // Fake success callback
-+        _albumCreated.value = true
-+    }
-+}
-diff --git a/app/src/main/res/layout/fragment_create_album.xml b/app/src/main/res/layout/fragment_create_album.xml
-index 66039d7..6d7166f 100644
---- a/app/src/main/res/layout/fragment_create_album.xml
-+++ b/app/src/main/res/layout/fragment_create_album.xml
-@@ -2,6 +2,7 @@
- <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-     android:layout_width="match_parent"
-     android:layout_height="match_parent"
-+    android:background="@color/white_background"
-     android:padding="16dp">
- 
-     <LinearLayout
-diff --git a/app/src/main/res/values/strings.xml b/app/src/main/res/values/strings.xml
-index 02e7be9..7afb112 100644
---- a/app/src/main/res/values/strings.xml
-+++ b/app/src/main/res/values/strings.xml
-@@ -1,3 +1,21 @@
- <resources>
-     <string name="app_name">VinilosMobileApp</string>
-+
-+    <!-- Home screen -->
-+    <string name="title_vinilos">Vinilos</string>
-+    <string name="title_artists">Artistas</string>
-+    <string name="title_favorites">Favoritos</string>
-+    <string name="title_profile">Perfil</string>
-+
-+    <!-- Album screen -->
-+    <string name="create_new_album">Crear Nuevo Álbum</string>
-+    <string name="detail_album">Detalles del Álbum</string>
-+
-+
-+    <!-- Artist screen -->
-+
-+
-+    <!-- Favorites screen -->
-+
-+
- </resources>
