+++ gradle/libs.versions.toml
@@ -9,11 +9,14 @@ lifecycleRuntimeKtx = "2.8.7"
 activityCompose = "1.10.1"
 composeBom = "2023.08.00"
 
-## Navigation
+## Other dependencies
 nav_version = "2.7.7"
+material_version = "1.11.0"
+material3_version = "1.1.0"
+glide_version = "4.16.0"
+coreKtxVersion = "1.6.1"
+junitKtx = "1.2.1"
 
-## Material
-material = "1.11.0"
 
 
 [libraries]
@@ -35,7 +38,11 @@ androidx-material3 = { group = "androidx.compose.material3", name = "material3" 
 ## Other dependencies
 navigation-fragment-ktx = { group = "androidx.navigation", name = "navigation-fragment-ktx", version.ref = "nav_version" }
 navigation-ui-ktx = { group = "androidx.navigation", name = "navigation-ui-ktx", version.ref = "nav_version" }
-material = { group = "com.google.android.material", name = "material", version.ref = "material" }
+material = { group = "com.google.android.material", name = "material", version.ref = "material_version" }
+androidx-material = { group = "androidx.compose.material3", name = "material3", version.ref = "material3_version" }
+glide = { group = "com.github.bumptech.glide", name = "glide", version.ref = "glide_version" }
+core-ktx = { group = "androidx.test", name = "core-ktx", version.ref = "coreKtxVersion" }
+androidx-junit-ktx = { group = "androidx.test.ext", name = "junit-ktx", version.ref = "junitKtx" }
 
 
 [plugins]
